refactor: enhance task definition update logic and add utility functions for testing
