+++ test/utils/__init__.py
@@ -1,0 +1,5 @@
+"""
+Test utilities module for Cloudlift.
+
+This module contains utility functions used across various test modules.
+"""
